+++ DiscoveryService/Messages/KademliaResponses/FindValueResponse.cs
@@ -47,7 +47,7 @@ namespace LUC.DiscoveryService.Messages.KademliaResponses
         {
             if (reader != null)
             {
-                MessageOperation = (MessageOperation)reader.ReadUInt32();
+                MessageOperation = (MessageOperation)reader.ReadUInt16();
                 RandomID = BigInteger.Parse(reader.ReadString());
                 CloseContactsToRepsonsingPeer = reader.ReadListOfContacts();
                 ValueInResponsingPeer = reader.ReadString();
