+++ DiscoveryService/Messages/KademliaResponses/Response.cs
@@ -18,7 +18,7 @@ namespace LUC.DiscoveryService.Messages.KademliaResponses
         {
             if (reader != null)
             {
-                MessageOperation = (MessageOperation)reader.ReadUInt32();
+                MessageOperation = (MessageOperation)reader.ReadUInt16();
                 RandomID = BigInteger.Parse(reader.ReadString());
 
                 return this;
